initial migration out of util
