+++ fence/config.py
@@ -1,4 +1,9 @@
 import os
+from functools import wraps
+
+import bcrypt
+import flask
+from werkzeug.datastructures import ImmutableMultiDict
 from yaml import safe_load as yaml_load
 import urllib.parse
 
@@ -6,6 +11,11 @@ import gen3cirrus
 from gen3config import Config
 
 from cdislogging import get_logger
+
+from fence import get_SQLAlchemyDriver
+from fence.utils import log_backoff_retry, log_backoff_giveup, exception_do_not_retry, generate_client_credentials, \
+    logger
+from fence.models import Client, User, query_for_user
 
 logger = get_logger(__name__)
 
@@ -169,3 +179,131 @@ class FenceConfig(Config):
 
 
 config = FenceConfig(DEFAULT_CFG_PATH)
+
+# Default settings to control usage of backoff library.
+DEFAULT_BACKOFF_SETTINGS = {
+    "on_backoff": log_backoff_retry,
+    "on_giveup": log_backoff_giveup,
+    "max_tries": config["DEFAULT_BACKOFF_SETTINGS_MAX_TRIES"],
+    "giveup": exception_do_not_retry,
+}
+
+
+def create_client(
+    DB,
+    username=None,
+    urls=[],
+    name="",
+    description="",
+    auto_approve=False,
+    is_admin=False,
+    grant_types=None,
+    confidential=True,
+    arborist=None,
+    policies=None,
+    allowed_scopes=None,
+    expires_in=None,
+):
+    client_id, client_secret, hashed_secret = generate_client_credentials(confidential)
+    if arborist is not None:
+        arborist.create_client(client_id, policies)
+    driver = get_SQLAlchemyDriver(DB)
+    auth_method = "client_secret_basic" if confidential else "none"
+
+    allowed_scopes = allowed_scopes or config["CLIENT_ALLOWED_SCOPES"]
+    if not set(allowed_scopes).issubset(set(config["CLIENT_ALLOWED_SCOPES"])):
+        raise ValueError(
+            "Each allowed scope must be one of: {}".format(
+                config["CLIENT_ALLOWED_SCOPES"]
+            )
+        )
+
+    if "openid" not in allowed_scopes:
+        allowed_scopes.append("openid")
+        logger.warning('Adding required "openid" scope to list of allowed scopes.')
+
+    with driver.session as s:
+        user = None
+        if username:
+            user = query_for_user(session=s, username=username)
+            if not user:
+                user = User(username=username, is_admin=is_admin)
+                s.add(user)
+
+        if s.query(Client).filter(Client.name == name).first():
+            if arborist is not None:
+                arborist.delete_client(client_id)
+            raise Exception("client {} already exists".format(name))
+
+        client = Client(
+            client_id=client_id,
+            client_secret=hashed_secret,
+            user=user,
+            redirect_uris=urls,
+            allowed_scopes=" ".join(allowed_scopes),
+            description=description,
+            name=name,
+            auto_approve=auto_approve,
+            grant_types=grant_types,
+            is_confidential=confidential,
+            token_endpoint_auth_method=auth_method,
+            expires_in=expires_in,
+        )
+        s.add(client)
+        s.commit()
+
+    return client_id, client_secret
+
+
+def hash_secret(f):
+    @wraps(f)
+    def wrapper(*args, **kwargs):
+        has_secret = "client_secret" in flask.request.form
+        has_client_id = "client_id" in flask.request.form
+        if flask.request.form and has_secret and has_client_id:
+            form = flask.request.form.to_dict()
+            with flask.current_app.db.session as session:
+                client = (
+                    session.query(Client)
+                    .filter(Client.client_id == form["client_id"])
+                    .first()
+                )
+                if client:
+                    form["client_secret"] = bcrypt.hashpw(
+                        form["client_secret"].encode("utf-8"),
+                        client.client_secret.encode("utf-8"),
+                    ).decode("utf-8")
+                flask.request.form = ImmutableMultiDict(form)
+
+        return f(*args, **kwargs)
+
+    return wrapper
+
+
+def get_from_cache(item_id, memory_cache, db_cache_table, db_cache_table_id_field="id"):
+    """
+    Attempt to get a cached item and store in memory cache from db if necessary.
+
+    NOTE: This requires custom implementation for putting items in the db cache table.
+    """
+    # try to retrieve from local in-memory cache
+    rv, expires_at = memory_cache.get(item_id, (None, 0))
+    if expires_at > expiry:
+        return rv
+
+    # try to retrieve from database cache
+    if hasattr(flask.current_app, "db"):  # we don't have db in startup
+        with flask.current_app.db.session as session:
+            cache = (
+                session.query(db_cache_table)
+                .filter(
+                    getattr(db_cache_table, db_cache_table_id_field, None) == item_id
+                )
+                .first()
+            )
+            if cache and cache.expires_at and cache.expires_at > expiry:
+                rv = dict(cache)
+
+                # store in memory cache
+                memory_cache[item_id] = rv, cache.expires_at
+                return rv
