+++ tests/scripting/test_fence-create.py
@@ -9,10 +9,10 @@ import gen3cirrus
 from gen3cirrus.google_cloud.errors import GoogleAuthError
 from userdatamodel.models import Group
 
-from fence.config import config
+from fence.config import config, create_client
 from fence.errors import UserError
 from fence.jwt.validate import validate_jwt
-from fence.utils import create_client, get_SQLAlchemyDriver
+from fence.utils import get_SQLAlchemyDriver
 from fence.models import (
     AccessPrivilege,
     Project,
